+++ src/os2datascanner/projects/admin/adminapp/views/webscanner_views.py
@@ -45,7 +45,7 @@ class WebScannerCreate(ScannerCreate):
               'download_sitemap', 'sitemap_url', 'sitemap', 'do_ocr',
               'do_link_check', 'only_notify_superadmin', 'do_last_modified_check',
               'rule', 'organization', 'exclude_urls', 'reduce_communication',
-              'keep_false_positives']
+              'keep_false_positives', 'always_crawl']
 
     def get_form(self, form_class=None):
         if form_class is None:
@@ -80,7 +80,7 @@ class WebScannerCopy(ScannerCopy):
               'download_sitemap', 'sitemap_url', 'sitemap', 'do_ocr',
               'do_link_check', 'only_notify_superadmin', 'do_last_modified_check',
               'rule', 'organization', 'exclude_urls', 'reduce_communication',
-              'keep_false_positives']
+              'keep_false_positives', 'always_crawl']
 
 
 class WebScannerUpdate(ScannerUpdate):
@@ -92,7 +92,7 @@ class WebScannerUpdate(ScannerUpdate):
               'download_sitemap', 'sitemap_url', 'sitemap', 'do_ocr',
               'do_link_check', 'only_notify_superadmin', 'do_last_modified_check',
               'rule', 'organization', 'exclude_urls', 'reduce_communication',
-              'keep_false_positives']
+              'keep_false_positives', 'always_crawl']
 
     def form_valid(self, form):
         if url_contains_spaces(form):